image-galoisry/requirements.txt
flask == 2.3.3
flask-session == 0.5.0
Pillow == 10.0.0
pymongo == 4.5.0
pycryptodome == 3.18.0
numpy == 1.25.2
apscheduler == 3.10.4
